+++ src/se_theory_reference_kit/declarations/__init__.py
@@ -2,12 +2,10 @@
 
 from se_theory_reference_kit.declarations.config import TheoryReferenceConfig
 from se_theory_reference_kit.declarations.export_spec import ExportSpec
-from se_theory_reference_kit.declarations.index import ReferenceIndex
 from se_theory_reference_kit.declarations.surface import SurfaceSymbols
 
 __all__ = [
     "ExportSpec",
-    "ReferenceIndex",
     "SurfaceSymbols",
     "TheoryReferenceConfig",
 ]